Tests added

+++ src/main/java/com/example/blabbercopy/event/SubscriptionChangeApplicationEvent.java
@@ -7,11 +7,11 @@ import org.springframework.context.ApplicationEvent;
 @Getter
 @ToString
 public class SubscriptionChangeApplicationEvent extends ApplicationEvent {
-    private int followeeId;
-    private int followerId;
+    private Long followeeId;
+    private Long followerId;
     private final SubscriptionType subscriptionType;
 
-    public SubscriptionChangeApplicationEvent(Object source, int foloweeId, int followerId, SubscriptionType subscriptionType) {
+    public SubscriptionChangeApplicationEvent(Object source, Long foloweeId, Long followerId, SubscriptionType subscriptionType) {
         super(source);
         this.followeeId = foloweeId;
         this.followerId = followerId;
